requirements.txt
joblib==1.1.0
matplotlib==3.5.2
numba==0.55.2
numpy==1.22.0
opencv_python==4.6.0.66
Pillow==9.3.0
PySide6==6.4.0.1
scikit_image==0.19.2
scikit_learn==1.1.3
skimage==0.0
torch==1.11.0+cu113
torchvision==0.12.0+cu113
